SAMI Subtitle sync, content Load
